Some default NixOS configuration per Quixoftic server requirements.
Note that the settings in config/defaults are enabled when you import
the module set. There are no "enable" flags. (They can, of course, be
overriden in a local deployment by using mkForce.)

For the most part these settings are passive and do not enable any
services. The exception is in config/defaults/system.nix, which
enables the logrotate service.