Add sudo tests; fix a bug in the sudo environment.
The tests caught a bug that has been lurking in my sudo settings for
probably quite awhile. In order to ensure that the user's TZ
environment variable isn't inherited in the sudo environment, it must
be removed not only from the env_keep list, but also the env_check
list (which seems to override env_keep). For good measure, we also set
"TZ=" in the sudo environment to be sure it's unset.