Add the first overlay (jemalloc on armv7l); fix multi-platform testing.
Only the tests know which systems they're being built for, and
makeTest needs to know that, too; so we can't import makeTest
generically in lib.nix as the system isn't known there at import
time. Therefore, I've moved the makeTest imports back into each test.

Note that armv7l-linux tests aren't working because nix-build is
trying to build numactl for that platform, and it doesn't appear to be
supported. Not sure yet why this is happening.