Overlays have moved to nixpkgs-quixoftic.
We now import that overlay set and use it instead of defining our own
overlays here. We do the import by fixing the rev and sha256 of the
nixpkgs-quixoftic GitHub repo we want to use, just like we do for
Nixpkgs. (And we have generalized fetch-nixpkgs.nix to fetch any
GitHub repo archive, so that we can use it for both nixpkgs and
nixpkgs-quixoftic.)

Note that this is just a crude direct port/replacement of the old
overlays. It might need a bit of a re-think about possible ways to do
this more elegantly or efficiently, but it works for now.