Use plaintext pw rather than hash in hydra-manual-setup service.
I can't get the hash to work; the password of the created user never
matches. The plaintext password works fine, however. This is less
secure, but not majorly so, since Hydra only uses SHA1 hashes anyway.

Also cleaned up the hydra-manual-setup service test, and added a test
to make sure the user is created correctly.