unbound-adblock: add support for multiple service IPs.
Also added IPv6 service support. However, the IPv6 tests fail. I
believe this has to do with the fact that there isn't proper routing
for the virtual service IPs on the server VM. In a typical routed
environment, you would run OSPF, static routing, or similar to route
these virtual IPv6's, so I don't think it's a problem with the module,
just the test. Anyway, the IPv6 tests are commented out for now.