hydra-manual-setup wants "keys.target" rather than specific keys.
This allows the user to use NixOps to specify a different key name in
/run/keys for the initial admin password and binary cache key and
still get the service to wait for the keys.

Previously we waited for the specific default key names, so if the
user wanted to override these, hydra-manual-setup would know to wait
for them.

Additionally, this is the canonical way to wait for keys in NixOS
modules (c.f. dovecot, nsd, strongswan).

Closes #3.