Move nvme and ahci modules on Intel to boot.initrd.kernelModules.
One of my Intel boxes uses an NVME device for its root filesystem. It
used to boot just fine with "nvme" listed in
boot.initrd.availableKernelModules (circa Dec 2017 nixpkgs), but with
a recent nixos-unstable-small channel version, it can no longer find
the NVME device at boot.

Moving "nvme" to boot.initrd.kernelModules fixes the problem. It's not
clear why, but this will do for now.

Note that as a precaution, to ensure that non-NVME hardware is using
AHCI for SATA devices, I've moved "ahci" into
boot.initrd.kernelModules, as well.