unbound-anycast: remove blocklist auto-updates.
I had second thoughts about this functionality. On the one hand,
there's a slight risk that an outdated adblock list will become so
stale that it'll be less useful. On the other hand, there's a slight
risk that the update script will fail, or worse, produce bogus data
that a malicious attacker could use to hijack your recursive DNS
server.

Lately the upstream block list seems to be updated only rarely, so
it's probably better to be safe here and just remove the ability to
update automatically. This also fits in better with the declarative
nature of NixOS: it should be clear from the derivation hash which
version of the blocklist is being used, since it's included with the
package.