Rename unbound-anycast -> unbound-multi-instance.
I think that requiring this service to be "anycast" is probably
unnecessary... not to mention the fact that the "anycast" module that
it relies upon doesn't really provide anycast support. The anycast
aspect will probably be quite site-specific and difficult to do
generically -- some will want to use OSPF, some RIP, some BGP,
etc.

So, instead let's just call this what it really is: a multi-instance
unbound service. The "listen addresses" can be any interface, and it's
up to the user to make sure those interfaces are set up.