ntp-server: remove listenIPs.
These don't really work as intended due to the design of ntpd. It's
not possible to configure ntpd to send upstream sync time requests on
a particular interface; it will choose one of the addresses it's using
for *listening* and use that interface for *sending*, as well. The
original intention of providing the `listenIPs` config option was to
prevent ntpd from listening on an external interface when it's running
on a firewall; but if you constraint ntpd to listen only on an
internal interface, and traffic from that interface isn't NATed, then
ntpd will be unable to reach upstream public Internet time servers.

The only real fix is to listen on all addresses, in which case ntpd
will figure out how to do the right thing to get upstream, and then
add firewall rules and ntpd application rules to permit just the
clients you want to allow to synchronize. (The former is out of scope
for this module; the latter is coming in a subsequent commit.)