Punt on a NTP server config and just add extraConfig to upstream.
There are too many ways to try to configure this, and too many ways to
get it wrong for this case or that one. So, let's just add an
`extraConfig` option to Nixpkgs services.ntp and leave it at
that. (Note that Nixpkgs should really have this, anyway, so let's
upstream this at some point.)